+++ demos/mandelbrot/src/distributed_canvas.h
@@ -6,99 +6,59 @@
 #include "mpicxx/mpi.h"
 
 #include "colours.h"
-#include "test/testutils.h"
 
 class distributed_canvas
 {
 public:
 
-    distributed_canvas(std::size_t width, std::size_t height, mpi::communicator comm = mpi::communicator::get_default())
-        : width_{width}, height_{height}, comm_{comm}
-    {
-        const auto total_size = local_height()*local_width();
-        data_ = std::vector<pixel>(total_size);
-    }
+    distributed_canvas(std::size_t width, std::size_t height, mpi::communicator comm = mpi::communicator::get_default());
 
     // Converts a local row index into its global index
-    std::size_t global_row(std::size_t local_row) const {
-        return local_height() * static_cast<std::size_t>(comm_.rank()) + local_row;
-    }
+    std::size_t global_row(std::size_t local_row) const;
 
     // Converts a local column index into its global index
-    std::size_t global_col(std::size_t local_col) const noexcept {
-        return local_col;
-    }
+    std::size_t global_col(std::size_t local_col) const noexcept;
 
     // Converts a global row index into its local index
-    std::size_t local_row(std::size_t g_row) const {
-        const std::size_t rowsbegin = global_row(0);
-        assert(g_row >= rowsbegin);
-        assert(g_row < global_row(local_height()));
-        return g_row - rowsbegin;
-    }
+    std::size_t local_row(std::size_t g_row) const ;
 
     // Converts a global column index into its local index
-    std::size_t local_col(std::size_t g_col) const noexcept {
-        const std::size_t colsbegin = global_col(0);
-        assert(g_col >= colsbegin);
-        assert(g_col < global_col(local_width()));
-        return g_col - colsbegin;
-    }
+    std::size_t local_col(std::size_t g_col) const noexcept;
 
     // Returs the width of the whole canvas
-    std::size_t global_width() const noexcept {
-        return width_;
-    }
+    std::size_t global_width() const noexcept;
 
     // Returs the width of this rank's section of the canvas
-    std::size_t local_width() const noexcept {
-        return width_;
-    }
+    std::size_t local_width() const noexcept;
 
     // Returs the heigh of the whole canvas
-    std::size_t global_height() const noexcept {
-        return height_;
-    }
+    std::size_t global_height() const noexcept;
 
     // Returs the height of this rank's section of the canvas
-    std::size_t local_height() const noexcept {
-        return height_ / static_cast<std::size_t>(comm_.size());
-    }
+    std::size_t local_height() const noexcept;
 
     // Range of pixel rows in this section
-    auto rows() const -> std::ranges::iota_view<std::size_t, std::size_t> {
-        return {global_row(0), global_row(local_height())};
-    }
+    auto rows() const -> std::ranges::iota_view<std::size_t, std::size_t>;
 
     // Range of pixel columns in this section
-    auto cols() const -> std::ranges::iota_view<std::size_t, std::size_t> {
-        return {global_col(0), global_col(local_width())};
-    }
+    auto cols() const -> std::ranges::iota_view<std::size_t, std::size_t>;
 
-    pixel& get(std::size_t g_row, std::size_t g_col) {
-        return data_[local_row(g_row)*local_width() + local_col(g_col)];
-    }
+    // Pixel located at certain global coordinates.
+    // Must be in this rank's section
+    pixel& get(std::size_t g_row, std::size_t g_col);
+    pixel const& get(std::size_t g_row, std::size_t g_col) const;
 
-    pixel const& get(std::size_t g_row, std::size_t g_col) const {
-        return data_[local_row(g_row)*local_width() + local_col(g_col)];
-    }
+    // Getter for comm_
+    mpi::communicator communicator() const;
 
-    mpi::communicator communicator() const {
-        return comm_;
-    }
-
-    std::span<pixel> flat_view() {
-        return {data_.begin(), data_.end()};
-    }
-
-    std::span<const pixel> flat_view() const {
-        return {data_.cbegin(), data_.cend()};
-    }
+    // Returns an iterable of all pixels
+    std::span<pixel> flat_view();
+    std::span<const pixel> flat_view() const;
    
 private:
-    std::size_t width_;
-    std::size_t height_;
-    mpi::communicator comm_;
-    std::vector<pixel> data_;
+    std::size_t width_;         // Global width of the canvas
+    std::size_t height_;        // Global height of the canvas
+    mpi::communicator comm_;    // Communicator to coordinate threads
+    std::vector<pixel> data_;   // Pixel data in this rank
 };
